mep fix

+++ controllers/comment.controller.js
@@ -1,5 +1,6 @@
 const Post = require("../database/models/post.model");
 const Comment = require("../database/models/comment.model");
+const User = require("../database/models/user.model");
 
 // get all existing comments from a post
 const getCommentsFromPost = async (req, res) => {
@@ -7,20 +8,17 @@ const getCommentsFromPost = async (req, res) => {
   const post = await Post.findOne({ postId });
 
   if (!post) {
-    console.log(`The post ${postId} does not exist !`);
     return res.status(404).json({ error: `This post does not exist !` });
   }
 
   const comments = await Comment.find({ postId });
 
   if (comments.length === 0) {
-    console.log(`No comments existed !`);
     return res
       .status(404)
       .json({ error: `No comments existed for this post !` });
   }
 
-  console.log(`List of comments : ${comments}`);
   return res.status(200).json({ response: comments });
 };
 
@@ -28,13 +26,17 @@ const getCommentsFromPost = async (req, res) => {
 const createCommentFromPost = async (req, res) => {
   const { postId } = req.params;
   const { content } = req.body;
-  const { userId } = req.session;
+  const { userId } = req;
 
   const post = await Post.findOne({ postId });
 
   if (!post) {
-    console.log(`The post ${postId} does not exist !`);
     return res.status(404).json({ error: `This post does not exist !` });
+  }
+
+  const isUserExist = await User.findOne({ userId });
+  if (!isUserExist) {
+    return res.status(500).json({ error: `User ${userId} does not exist !` });
   }
 
   const comment = new Comment({
@@ -50,21 +52,12 @@ const createCommentFromPost = async (req, res) => {
       post.markModified("comments");
       await post.save();
 
-      console.log(
-        `New comment ${comment.commentId} has been successfully created for post ${postId} by user ${userId} !`
-      );
-
-      console.log(
-        `The comment ${comment.commentId} from the post ${postId} has been successfully created by user ${userId}`
-      );
-
       return res
         .status(201)
         .json({ response: `Your comment has been successfully created !` });
     })
     .catch((err) => {
       if (err) {
-        console.log(`Error comment creation for user ${userId} :`, err);
         return res.status(500).json({ error: err });
       }
     });
@@ -74,7 +67,7 @@ const createCommentFromPost = async (req, res) => {
 const updateCommentFromPost = async (req, res) => {
   const { postId, commentId } = req.params;
   const { content } = req.body;
-  const { userId } = req.session;
+  const { userId } = req;
 
   if (!content) {
     return res.status(400).json({ error: `Please fill in all fields !` });
@@ -87,10 +80,12 @@ const updateCommentFromPost = async (req, res) => {
     return res.status(404).json({ error: `Comment does not exist !` });
   }
 
+  const isUserExist = await User.findOne({ userId });
+  if (!isUserExist) {
+    return res.status(500).json({ error: `User ${userId} does not exist !` });
+  }
+
   if (userId !== comment.userId) {
-    console.log(
-      `User ${userId} wants to modify the comment ${commentId} but it was created by user ${comment.userId} !`
-    );
     return res.status(403).json({
       error: `Permission denied, you cannot edit the comment of another user !`,
     });
@@ -102,23 +97,12 @@ const updateCommentFromPost = async (req, res) => {
   const post = await Post.findOne({ postId });
 
   if (!post) {
-    console.log(
-      `User ${userId} has problem with editing comment ${commentId} of the post ${postId} !`
-    );
     return res.status(404).json({ error: `Problem while editing comment !` });
   }
 
   post.comments[commentId] = content;
   post.markModified("comments");
   await post.save();
-
-  console.log(
-    `Comment ${commentId} has been successfully updated by ${userId} !`
-  );
-
-  console.log(
-    `The comment ${commentId} from the post ${postId} has been successfully updated by user ${userId}`
-  );
 
   return res
     .status(200)
@@ -128,22 +112,23 @@ const updateCommentFromPost = async (req, res) => {
 // delete one comment from an existing post
 const deleteCommentFromPost = async (req, res) => {
   const { postId, commentId } = req.params;
-  const { userId } = req.session;
+  const { userId } = req;
 
   const isCommentExist = await Comment.findOne({ commentId });
 
   if (!isCommentExist) {
-    console.log(`Comment ${commentId} does not exist !`);
     return res.status(404).json({ error: `Comment does not exist !` });
   }
 
   if (isCommentExist.userId !== userId) {
-    console.log(
-      `User ${userId} wants to delete the comment ${commentId} but it was created by user ${isCommentExist.userId} !`
-    );
     return res.status(403).json({
       error: `Permission denied, you cannot delete the comment of another user !`,
     });
+  }
+
+  const isUserExist = await User.findOne({ userId });
+  if (!isUserExist) {
+    return res.status(500).json({ error: `User ${userId} does not exist !` });
   }
 
   const comment = await Comment.deleteOne({ commentId });
@@ -151,23 +136,12 @@ const deleteCommentFromPost = async (req, res) => {
   const post = await Post.findOne({ postId });
 
   if (!post) {
-    console.log(
-      `User ${userId} has problem with deleting comment ${commentId} of the post ${postId} !`
-    );
     return res.status(404).json({ error: `Problem while deleting comment !` });
   }
 
   delete post.comments[commentId];
   post.markModified("comments");
   await post.save();
-
-  console.log(
-    `Comment ${commentId} has been successfully deleted by ${userId} !`
-  );
-
-  console.log(
-    `The comment ${commentId} from the post ${postId} has been successfully deleted by user ${userId} !`
-  );
 
   return res
     .status(202)
@@ -179,20 +153,20 @@ const createLikeCommentFromPost = async (req, res) => {
   const { commentId } = req.params;
   const { userId } = req.body;
 
-  if (userId !== req.session.userId) {
-    console.log(
-      `Error occured, user ${req.session.userId} cannot like the comment ${commentId} !`
-    );
+  if (userId !== req.userId) {
     return res
       .status(400)
       .json({ err: `Error occured, you cannot like the comment !` });
   }
 
+  const isUserExist = await User.findOne({ userId });
+  if (!isUserExist) {
+    return res.status(500).json({ error: `User ${userId} does not exist !` });
+  }
+
   const comment = await Comment.findOne({ commentId });
 
   if (comment.likes.includes(userId)) {
-    console.log(`User ${userId} has already liked the comment ${commentId}`);
-
     return res.status(500).json({ error: `You already liked the comment !` });
   }
 
@@ -201,20 +175,11 @@ const createLikeCommentFromPost = async (req, res) => {
   await comment
     .save()
     .then(() => {
-      console.log(
-        `User ${userId} has successfully liked the comment ${commentId}`
-      );
-
       return res
         .status(200)
         .json({ response: `You successfully liked the comment !` });
     })
-    .catch((err) => {
-      console.log(
-        `Error occured with the like of the comment ${commentId} !`,
-        err
-      );
-
+    .catch(() => {
       return res
         .status(400)
         .json({ error: `Error occured with the like of the comment !` });
@@ -226,10 +191,12 @@ const deleteLikeCommentFromPost = async (req, res) => {
   const { commentId } = req.params;
   const { userId } = req.body;
 
-  if (userId !== req.session.userId) {
-    console.log(
-      `Error occured, user ${req.session.userId} cannot unlike the comment ${commentId} !`
-    );
+  const isUserExist = await User.findOne({ userId });
+  if (!isUserExist) {
+    return res.status(500).json({ error: `User ${userId} does not exist !` });
+  }
+
+  if (userId !== req.userId) {
     return res
       .status(400)
       .json({ err: `Error occured, you cannot unlike the comment !` });
@@ -238,8 +205,6 @@ const deleteLikeCommentFromPost = async (req, res) => {
   const comment = await Comment.findOne({ commentId });
 
   if (!comment.likes.includes(userId)) {
-    console.log(`User ${userId} cannot unlike the comment ${commentId} !`);
-
     return res
       .status(500)
       .json({ error: `You cannot unlike a comment you have not liked !` });
@@ -254,20 +219,11 @@ const deleteLikeCommentFromPost = async (req, res) => {
   await comment
     .save()
     .then(() => {
-      console.log(
-        `User ${userId} has successfully unliked the comment ${commentId}`
-      );
-
       return res
         .status(200)
         .json({ response: `You successfully unliked the comment !` });
     })
-    .catch((err) => {
-      console.log(
-        `Error occured, user ${userId} cannot unlike the comment ${commentId} !`,
-        err
-      );
-
+    .catch(() => {
       return res
         .status(400)
         .json({ error: `Error occured, you cannot unlike the comment !` });
