+++ controllers/post.controller.js
@@ -1,26 +1,30 @@
 const Post = require("../database/models/post.model");
 const Comment = require("../database/models/comment.model");
+const User = require("../database/models/user.model");
 
 // get all existing posts
 const getPosts = async (req, res) => {
   const posts = await Post.find({});
 
   if (posts.length === 0) {
-    console.log(`No posts existed !`);
     return res.status(404).json({ error: `No posts existed !` });
   }
 
-  console.log(`List of posts : ${posts}`);
   return res.status(200).json({ response: posts });
 };
 
 // create one post if it does not exist
 const createPost = async (req, res) => {
   const { topic, title, content } = req.body;
-  const userId = req.session.userId;
+  const { userId } = req;
 
   if (!userId || !topic || !title || !content) {
     return res.status(400).json({ error: `Please fill in all fields !` });
+  }
+
+  const isUserExist = await User.findOne({ userId });
+  if (!isUserExist) {
+    return res.status(500).json({ error: `User ${userId} does not exist !` });
   }
 
   const post = new Post({
@@ -34,17 +38,12 @@ const createPost = async (req, res) => {
   await post
     .save()
     .then(async () => {
-      console.log(
-        `New post ${post.postId} has been successfully created by user ${userId} !`
-      );
-
       return res
         .status(201)
         .json({ response: `Your post has been successfully created !` });
     })
     .catch((err) => {
       if (err) {
-        console.log(`Error post creation :`, err);
         return res.status(500).json({ error: err });
       }
     });
@@ -56,11 +55,9 @@ const getPost = async (req, res) => {
   const post = await Post.findOne({ postId });
 
   if (!post) {
-    console.log(`Post ${postId} does not exist !`);
     return res.status(404).json({ error: `Post ${postId} does not exist !` });
   }
 
-  console.log(`Post ${postId} :`, post);
   return res.status(200).json({ post });
 };
 
@@ -68,25 +65,24 @@ const getPost = async (req, res) => {
 const updatePost = async (req, res) => {
   const { postId } = req.params;
   const { topic, title, content } = req.body;
-  const { userId } = req.session;
+  const { userId } = req;
 
   if (!topic || !title || !content) {
     return res.status(400).json({ error: `Please fill in all fields !` });
   }
 
+  const isUserExist = await User.findOne({ userId });
+  if (!isUserExist) {
+    return res.status(500).json({ error: `User ${userId} does not exist !` });
+  }
+
   const post = await Post.findOne({ postId });
 
   if (!post) {
-    console.log(
-      `The user ${userId} wants to update the post ${postId} but does not exist !`
-    );
     return res.status(404).json({ error: `The post does not exist !` });
   }
 
   if (post.userId !== userId) {
-    console.log(
-      `User ${userId} wants to update the post ${postId} but it was created by user ${post.userId} !`
-    );
     return res.status(403).json({
       error: `Permission denied, you cannot update the post of another user !`,
     });
@@ -98,10 +94,6 @@ const updatePost = async (req, res) => {
 
   await post.save();
 
-  console.log(
-    `The post ${postId} has been successfully updated by user ${userId}`
-  );
-
   return res
     .status(200)
     .json({ response: `Post has been successfully updated !` });
@@ -110,32 +102,28 @@ const updatePost = async (req, res) => {
 // delete one post if it exists
 const deletePost = async (req, res) => {
   const { postId } = req.params;
-  const { userId } = req.session;
+  const { userId } = req;
+
   const isPostExist = await Post.findOne({ postId });
 
   if (!isPostExist) {
-    console.log(`Post ${postId} does not exist !`);
     return res.status(404).json({ error: `Post ${postId} does not exist !` });
   }
 
+  const isUserExist = await User.findOne({ userId });
+  if (!isUserExist) {
+    return res.status(500).json({ error: `User ${userId} does not exist !` });
+  }
+
   if (isPostExist.userId !== userId) {
-    console.log(
-      `User ${userId} wants to delete the post ${postId} but it was created by user ${isPostExist.userId} !`
-    );
     return res.status(403).json({
       error: `Permission denied, you cannot delete the post of another user !`,
     });
   }
 
   const post = await Post.deleteOne({ postId })
-    .then(() => {
-      console.log(
-        `Post ${postId} and its comments have been successfully deleted !`
-      );
-    })
+    .then(() => {})
     .catch((err) => {
-      console.log(`Problem while deleting post !`, err);
-
       return res
         .status(500)
         .json({ error: `Problem while deleting post !`, err });
@@ -143,17 +131,11 @@ const deletePost = async (req, res) => {
 
   const comments = await Comment.deleteMany({ postId })
     .then(() => {
-      console.log(
-        `All comments from the post ${postId} have been successfully deleted !`
-      );
-
       return res
         .status(202)
         .json({ response: `The post has been successfully deleted !` });
     })
     .catch((err) => {
-      console.log(`Problem while deleting post !`, err);
-
       return res
         .status(500)
         .json({ error: `Problem while deleting post !`, err });
@@ -165,10 +147,12 @@ const createLikePost = async (req, res) => {
   const { postId } = req.params;
   const { userId } = req.body;
 
-  if (userId !== req.session.userId) {
-    console.log(
-      `Error occured, user ${req.session.userId} cannot like the post ${postId} !`
-    );
+  const isUserExist = await User.findOne({ userId });
+  if (!isUserExist) {
+    return res.status(500).json({ error: `User ${userId} does not exist !` });
+  }
+
+  if (userId !== req.userId) {
     return res
       .status(400)
       .json({ err: `Error occured, you cannot like the post !` });
@@ -177,8 +161,6 @@ const createLikePost = async (req, res) => {
   const post = await Post.findOne({ postId });
 
   if (post.likes.includes(userId)) {
-    console.log(`User ${userId} has already liked the post ${postId}`);
-
     return res.status(500).json({ error: `You already liked the post !` });
   }
 
@@ -187,15 +169,11 @@ const createLikePost = async (req, res) => {
   await post
     .save()
     .then(() => {
-      console.log(`User ${userId} has successfully liked the post ${postId}`);
-
       return res
         .status(200)
         .json({ response: `You successfully liked the post !` });
     })
     .catch((err) => {
-      console.log(`Error occured with the like of the post ${postId} !`, err);
-
       return res
         .status(400)
         .json({ error: `Error occured with the like of the post !` });
@@ -207,10 +185,12 @@ const deleteLikePost = async (req, res) => {
   const { postId } = req.params;
   const { userId } = req.body;
 
-  if (userId !== req.session.userId) {
-    console.log(
-      `Error occured, user ${req.session.userId} cannot unlike the post ${postId} !`
-    );
+  const isUserExist = await User.findOne({ userId });
+  if (!isUserExist) {
+    return res.status(500).json({ error: `User ${userId} does not exist !` });
+  }
+
+  if (userId !== req.userId) {
     return res
       .status(400)
       .json({ err: `Error occured, you cannot unlike the post !` });
@@ -219,8 +199,6 @@ const deleteLikePost = async (req, res) => {
   const post = await Post.findOne({ postId });
 
   if (!post.likes.includes(userId)) {
-    console.log(`User ${userId} cannot unlike the post ${postId} !`);
-
     return res
       .status(500)
       .json({ error: `You cannot unlike a post you have not liked !` });
@@ -235,18 +213,11 @@ const deleteLikePost = async (req, res) => {
   await post
     .save()
     .then(() => {
-      console.log(`User ${userId} has successfully unliked the post ${postId}`);
-
       return res
         .status(200)
         .json({ response: `You successfully unliked the post !` });
     })
     .catch((err) => {
-      console.log(
-        `Error occured, user ${userId} cannot unlike the post ${postId} !`,
-        err
-      );
-
       return res
         .status(400)
         .json({ error: `Error occured, you cannot unlike the post !` });
